requirements.txt
python-dotenv==1.0.1
pyyaml==5.3.1
openai==1.14.2
requests==2.31.0
redis==5.0.4

